refactor(common): move `DummyApiVersionDiscriminantSource` to `test::double`
