+++ mithril-common/src/api_version.rs
@@ -73,38 +73,10 @@ impl APIVersionProvider {
     }
 }
 
-cfg_test_tools! {
-    /// A dummy implementation of the `ApiVersionDiscriminantSource` trait for testing purposes.
-    pub struct DummyApiVersionDiscriminantSource {
-        discriminant: String,
-    }
-
-    impl DummyApiVersionDiscriminantSource {
-        /// Create a new instance of `DummyApiVersionDiscriminantSource` with the given discriminant.
-        pub fn new<T: Into<String>>(discrimant: T) -> Self {
-            Self {
-                discriminant: discrimant.into(),
-            }
-        }
-    }
-
-    impl Default for DummyApiVersionDiscriminantSource {
-        fn default() -> Self {
-            Self {
-                discriminant: "dummy".to_string(),
-            }
-        }
-    }
-
-    impl ApiVersionDiscriminantSource for DummyApiVersionDiscriminantSource {
-        fn get_discriminant(&self) -> String {
-            self.discriminant.clone()
-        }
-    }
-}
-
 #[cfg(test)]
 mod test {
+    use crate::test::double::DummyApiVersionDiscriminantSource;
+
     use super::*;
 
     #[test]
